altera nomes e conserta erros de valores nulos

+++ src/main/java/br/ufrn/imd/incluevents/ufpb/dto/CreateUsuarioDtoUfpb.java
@@ -4,6 +4,6 @@ import br.ufrn.imd.incluevents.framework.dto.CreateUsuarioDto;
 import br.ufrn.imd.incluevents.ufpb.model.enums.CargoEnumUfpb;
 import br.ufrn.imd.incluevents.ufpb.model.enums.TipoUsuarioEnumUfpb;
 
-public record CreateUsuarioDtoUfpb(String nome, String email, String senha, String username, TipoUsuarioEnumUfpb tipo, CargoEnumUfpb cargo, Integer tempoServico, Double imc) implements CreateUsuarioDto {
+public record CreateUsuarioDtoUfpb(String nome, String email, String senha, String username, TipoUsuarioEnumUfpb tipo, CargoEnumUfpb cargo, Double tempoServico, Double imc) implements CreateUsuarioDto {
 
 }
